+++ utils/kafka-utils/src/main/scala/pl/touk/nussknacker/engine/kafka/exception/KafkaErrorTopicInitializer.scala
@@ -1,39 +1,46 @@
 package pl.touk.nussknacker.engine.kafka.exception
 
 import com.typesafe.scalalogging.LazyLogging
-import org.apache.kafka.clients.admin.{ListTopicsOptions, NewTopic}
+import org.apache.kafka.clients.admin.{DescribeTopicsOptions, ListTopicsOptions, NewTopic}
 import pl.touk.nussknacker.engine.kafka.{KafkaConfig, KafkaUtils}
 
-import java.util.Optional
+import java.util.{Collections, Optional}
 import java.util.concurrent.TimeUnit
 import java.{lang, util}
 import scala.jdk.CollectionConverters._
 import scala.util.control.NonFatal
 
-class KafkaErrorTopicInitializer(kafkaConfig: KafkaConfig, exceptionHandlerConfig: KafkaExceptionConsumerConfig)
-    extends LazyLogging {
+trait KafkaErrorTopicInitializer {
+  def topicName: String
+  def init(): Unit
+}
+
+class DefaultKafkaErrorTopicInitializer(kafkaConfig: KafkaConfig, exceptionHandlerConfig: KafkaExceptionConsumerConfig)
+    extends KafkaErrorTopicInitializer
+    with LazyLogging {
 
   private val timeoutSeconds = 5
 
+  val topicName: String = exceptionHandlerConfig.topic
+
   def init(): Unit = {
-    val errorTopic = exceptionHandlerConfig.topic
     KafkaUtils.usingAdminClient(kafkaConfig) { admin =>
       val topicNames = admin
         .listTopics(new ListTopicsOptions().timeoutMs(timeoutSeconds * 1000))
         .names()
         .get(timeoutSeconds, TimeUnit.SECONDS)
-      val topicExists = topicNames.asScala.contains(errorTopic)
+      val topicExists = topicNames.asScala.contains(topicName)
       if (topicExists) {
-        logger.debug("Topic exists, skipping")
+        logger.debug(s"Topic $topicName already exists, skipping")
       } else {
-        logger.info(s"Creating error topic: $errorTopic with default configs, please check if the values are correct")
-        val errorTopicConfig = new NewTopic(errorTopic, Optional.empty[Integer](), Optional.empty[lang.Short]())
+        logger.info(s"Creating error topic: $topicName with default configs, please check if the values are correct")
+        val errorTopicConfig = new NewTopic(topicName, Optional.empty[Integer](), Optional.empty[lang.Short]())
         try {
           admin.createTopics(util.Arrays.asList(errorTopicConfig)).all().get(timeoutSeconds, TimeUnit.SECONDS)
         } catch {
           case NonFatal(e) =>
             throw new IllegalStateException(
-              s"Failed to create $errorTopic (${e.getMessage}), cannot run scenario properly",
+              s"Failed to create $topicName (${e.getMessage}), cannot run scenario properly",
               e
             )
         }
@@ -41,4 +48,11 @@ class KafkaErrorTopicInitializer(kafkaConfig: KafkaConfig, exceptionHandlerConfi
     }
   }
 
+}
+
+object NoopKafkaErrorTopicInitializer extends KafkaErrorTopicInitializer {
+
+  override val topicName: String = "-"
+
+  override def init(): Unit = {}
 }